Fake Players shoot at pretty much anything they see.
Main issue is the lack of pitch when they aim.
This is due to the anim blueprint using the controller remote pitch value, which would first of all read from the GDK Player Controller, not the non-rep'd SimPlayer controller, and secondly the Controller SetFocus method only controls yaw, not pitch.
This is something we will need to fix for AI anyway, so suggest including it as is for now.