Add Calls to Parent BeginPlay in Character Classes (#46)
In AGDKCharacter::BeginPlay a callback is registered on the UEquippedComponent::HoldableUpdated. This callback is used to attach the new AHoldable to the player's skeleton.

Blueprint subclasses: BP_Base_Character, and BP_FPS_Character had overridden BeginPlay but not called the Parent function, so the callback was not attached.

This meant that if a weapon was checked out after the character, it would not be attached, and would stay floating where it was spawned.